refactor: create core/constants.ts, migrate all imports, delete old constants.ts
