+++ src/admin/constants.ts
@@ -1,4 +1,4 @@
-// HTTP 状态码统一从 src/constants.ts 导入，避免重复定义
+// HTTP 状态码统一从 core/constants.ts 导入，避免重复定义
 export {
   HTTP_BAD_REQUEST,
   HTTP_CREATED,
@@ -8,4 +8,4 @@ export {
   HTTP_INTERNAL_ERROR,
   HTTP_BAD_GATEWAY,
   HTTP_SERVICE_UNAVAILABLE,
-} from "../constants.js";
+} from "../core/constants.js";
