--- src/constants.ts
@@ -1,7 +1,0 @@
-// src/constants.ts — 向后兼容 re-export，迁移完毕后删除
-export {
-  HTTP_BAD_REQUEST, HTTP_CREATED, HTTP_FORBIDDEN, HTTP_NOT_FOUND,
-  HTTP_CONFLICT, HTTP_INTERNAL_ERROR, HTTP_BAD_GATEWAY,
-  HTTP_UNPROCESSABLE_ENTITY, HTTP_SERVICE_UNAVAILABLE,
-  PROXY_API_TYPES, getProxyApiType, MS_PER_SECOND,
-} from "./core/constants.js";
