+++ src/core/constants.ts
@@ -1,3 +1,5 @@
+// src/core/constants.ts
+
 // HTTP 状态码常量 — 全局唯一来源
 export const HTTP_BAD_REQUEST = 400;
 export const HTTP_CREATED = 201;
