+++ src/proxy/openai.ts
@@ -15,7 +15,7 @@ import { ProviderSemaphoreManager } from "./semaphore.js";
 import type { RequestTracker } from "../monitor/request-tracker.js";
 import type { UsageWindowTracker } from "./usage-window-tracker.js";
 import type { AdaptiveConcurrencyController } from "./adaptive-controller.js";
-import { HTTP_NOT_FOUND, HTTP_BAD_GATEWAY } from "../constants.js";
+import { HTTP_NOT_FOUND, HTTP_BAD_GATEWAY } from "../core/constants.js";
 
 export interface OpenaiProxyOptions {
   db: Database.Database;
