+++ src/proxy/proxy-logging.ts
@@ -12,7 +12,7 @@ import type { TransportResult } from "./types.js";
 // Internal imports from types.ts
 import { UPSTREAM_SUCCESS } from "./types.js";
 import type { RawHeaders } from "./types.js";
-import { HTTP_BAD_GATEWAY } from "../constants.js";
+import { HTTP_BAD_GATEWAY } from "../core/constants.js";
 
 // ---------- Header sanitization ----------
