+++ src/proxy/resilience.ts
@@ -1,4 +1,4 @@
-import { MS_PER_SECOND } from "../constants.js";
+import { MS_PER_SECOND } from "../core/constants.js";
 import type { RetryRuleMatcher } from "./retry-rules.js";
 import { ProviderSwitchNeeded } from "./types.js";
 import type { TransportResult } from "./types.js";
